fix delete category but not refreshable yet
