+++ src/app/(admin)/products/page.tsx
@@ -1,9 +1,7 @@
 "use client";
-import { useState } from "react";
 import { 
   Card, Dropdown, DropdownItem, Button, 
-  Badge, TextInput, Select, 
-  Label, Modal, Textarea, 
+  Badge, TextInput, Select, Label, 
 } from "flowbite-react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { 
@@ -17,24 +15,9 @@ import { useSession } from "next-auth/react";
 import CustomSpinner from "@/components/CustomSpinner";
 import { useSearchSortPaginationStore } from "@/store/useSearchSortPaginationStore";
 import { useDebouncedCallback } from "use-debounce";
-import { ErrorMessage, Form, Formik, Field } from "formik";
-import * as Yup from "yup";
-import { CreateCategoryRequest, UpdateCategoryRequest } from "@/types/category";
-import { useSidebarAdminStore } from "@/store/useSidebarAdminStore";
-import ConfirmationModal from "@/components/ConfirmationModal";
-import useCreateCategory from "@/hooks/category/useCreateCategory";
-import useUpdateCategory from "@/hooks/category/useUpdateCategory";
-
-const validationSchema = Yup.object({
-  name: Yup.string()
-    .required("Name is required"),
-});
+import Category from "./components/Category";
 
 const Products = () => {
-  const [openModalConfirmation, setOpenModalConfirmation] = useState<boolean>(false);
-  const { 
-    isModalCategoryOpen, setIsModalCategoryOpen, modalCategoryType, category, 
-  } = useSidebarAdminStore();
   const { data: session } = useSession();
   const { 
     page, length, search, field, order, 
@@ -46,8 +29,6 @@ const Products = () => {
     products,
     totalData,
   } = useAdminProduct(session?.accessToken as string);
-  const { createCategory } = useCreateCategory(session?.accessToken as string);
-  const { updateCategory } = useUpdateCategory(session?.accessToken as string);
 
   if (productError) {
     return <div className="align-middle justify-center">Error: {productError.message}</div>;
@@ -65,23 +46,6 @@ const Products = () => {
 
   const handlePageChange = (newPage: number) => {
     setPage(newPage);
-  };
-
-  const handleSubmit = async (
-    values: CreateCategoryRequest | UpdateCategoryRequest
-  ) => {
-      try {
-        if(modalCategoryType === "create") {
-          createCategory({creationData: values as CreateCategoryRequest});
-        } else if(modalCategoryType === "update") {
-          updateCategory({
-            id: category.id.toString(), updateData: values as UpdateCategoryRequest
-          });
-        }
-      } catch (error) {
-        console.error("An unexpected error occurred:", error);
-      } finally {
-    }
   };
 
   if (isProductLoading) {
@@ -184,61 +148,7 @@ const Products = () => {
           Next
         </Button>
       </div>
-
-      <Modal show={isModalCategoryOpen} onClose={() => setIsModalCategoryOpen(false)}>
-      <Formik<CreateCategoryRequest | UpdateCategoryRequest>
-        initialValues={modalCategoryType == "update" ? category : { name: "" }}
-        validationSchema={validationSchema}
-        onSubmit={()=>{
-          setOpenModalConfirmation(true);
-        }}
-      >
-        {({ setSubmitting, isSubmitting, values }) => (
-          <Form className="flex flex-col w-full">
-            <Modal.Header>
-              <span className="capitalize">{modalCategoryType}</span>
-              &nbsp;Category
-            </Modal.Header>
-            <Modal.Body>
-              <div>
-                <Label htmlFor="name" className="font-medium">Category Name</Label>
-                <Field as={TextInput} id="name" name="name" placeholder="Enter category name" />
-                <ErrorMessage
-                  name="name"
-                  component="div"
-                  className="text-red-500 text-sm"
-                />
-              </div>
-              <div>
-                <Label htmlFor="description" className="font-medium">Description</Label>
-                <Textarea id="description" name="description" 
-                  placeholder="Enter category description" rows={3} />
-              </div>
-            </Modal.Body>
-            <Modal.Footer className="flex justify-between bg-ghost-white">
-              <Button color="gray" onClick={() => {
-                setIsModalCategoryOpen(false);
-              }} disabled={isSubmitting}>Cancel</Button>
-              <Button color="blue" disabled={isSubmitting} type="submit">
-                <span className="capitalize">{modalCategoryType}</span>
-              </Button>
-            </Modal.Footer>
-
-            <ConfirmationModal 
-              isOpen={openModalConfirmation}
-              onClose={()=>{
-                setSubmitting(false);
-                setOpenModalConfirmation(false);
-              }}
-              onConfirm={()=>{
-                handleSubmit(values);
-                setOpenModalConfirmation(false);
-              }}
-            />
-          </Form>
-        )}
-        </Formik>
-      </Modal>
+      <Category />
     </div>
   );
 };
