+++ src/store/useSidebarAdminStore.ts
@@ -7,11 +7,13 @@ export interface SidebarAdminState {
   isModalCategoryOpen: boolean;
   modalCategoryType: string;
   category: UpdateCategoryRequest;
+  isDeletingCategory: boolean;
   setIsOpen: (isOpen: boolean) => void;
   setPage: (page: string) => void;
   setIsModalCategoryOpen: (isModalCategoryOpen: boolean) => void;
   setModalCategoryType: (page: string) => void;
   setCategory: (category: UpdateCategoryRequest) => void;
+  setIsDeletingCategory: (isDeletingCategory: boolean) => void;
 }
 
 export const useSidebarAdminStore = create<SidebarAdminState>((set) => ({
@@ -20,9 +22,11 @@ export const useSidebarAdminStore = create<SidebarAdminState>((set) => ({
   isModalCategoryOpen: false,
   modalCategoryType: "create",
   category: {} as UpdateCategoryRequest,
+  isDeletingCategory: false, 
   setIsOpen: (isOpen) => set({ isOpen }),
   setPage: (page) => set({ page }),
   setIsModalCategoryOpen: (isModalCategoryOpen) => set({ isModalCategoryOpen }),
   setModalCategoryType: (modalCategoryType) => set({ modalCategoryType }),
   setCategory: (category) => set({ category }),
+  setIsDeletingCategory: (isDeletingCategory) => set({ isDeletingCategory }),
 }));
